Solved part 2 of day20 in 2019, part 2 adds one dimension called level to position in maze and defines a recursive about warping in maze, what a fantastic design! Let me recall the movie called inception. All changes in code is about new position definition, and rules about gate state(on/off) and how to warp between those gates. I used trait maze to abstract interface between walker and maze, this makes the searching code almost unchanged.
